Minor fix for Dart SDK `HttpRequest` and `HttpClientResponse` breakage
Newer versions of the Dart SDK (2.5.0-dev.1.0) has introduced API breakage,
resulting in the following build failure:

lib/handlers.dart:10:45: Error: The argument type 'Utf8Decoder' can't be assigned to the parameter type 'StreamTransformer<Uint8List, dynamic>'.
 - 'Utf8Decoder' is from 'dart:convert'.
 - 'StreamTransformer' is from 'dart:async'.
 - 'Uint8List' is from 'dart:typed_data'.
Try changing the type of the parameter, or casting the argument to 'StreamTransformer<Uint8List, dynamic>'.
  String content = await req.transform(utf8.decoder).join();
                                            ^

We fix this up in a way that should work for both older and newer
versions of the SDK.